requirements.txt
# requirements.txt
Pillow
numpy
tqdm
scikit-image